Move service around
